config.py
import os
from dotenv import load_dotenv
OLLAMA_MODEL = "llama3.1"

# Load environment variables from .env file
load_dotenv()

EMAIL_SENDER = os.getenv("EMAIL_SENDER")
EMAIL_PASSWORD = os.getenv("EMAIL_PASSWORD")
SMTP_USER = os.getenv("SMTP_USER")
SMTP_PASSWORD = os.getenv("SMTP_PASSWORD")
SMTP_SERVER = os.getenv("SMTP_SERVER")
SMTP_PORT = int(os.getenv("SMTP_PORT", 587))  
DB_PATH = os.getenv("DB_PATH", "recruitment.db") 

# Print to verify
print(f"Loaded EMAIL_SENDER: {EMAIL_SENDER}")
