requirements.txt
sqlite3
pandas
python-docx
PyMuPDF
ollama
smtplib
email
python-dotenv
